# specs/001-question-practice/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: YoungCerti Question API
  version: 1.0.0
  description: |
    Read-only REST API for the YoungCerti question practice MVP.
    Backed by an in-memory pool loaded from S3. No per-user state;
    no writes other than potential future operational endpoints.
servers:
  - url: https://api.young-certi.example/api/v1
    description: Production (CloudFront-fronted ALB to be confirmed at deploy time)
  - url: http://localhost:8000/api/v1
    description: Local docker-compose
paths:
  /healthz:
    get:
      summary: Liveness probe
      operationId: getHealthz
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /readyz:
    get:
      summary: Readiness probe (pool loaded)
      operationId: getReadyz
      responses:
        "200":
          description: Pool is loaded; ready to serve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: Pool not yet loaded or last refresh failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /exams:
    get:
      summary: List available exams
      operationId: listExams
      responses:
        "200":
          description: Exam list (only `sap-c02` in MVP)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamSummary"
  /exams/{examSlug}/questions:
    parameters:
      - $ref: "#/components/parameters/ExamSlug"
    get:
      summary: List question numbers for an exam
      operationId: listQuestionNumbers
      responses:
        "200":
          description: Ordered list of question numbers
          content:
            application/json:
              schema:
                type: object
                required: [examSlug, version, total, numbers]
                properties:
                  examSlug:
                    type: string
                    example: sap-c02
                  version:
                    type: string
                    example: v2025-11-26
                  total:
                    type: integer
                    example: 476
                  numbers:
                    type: array
                    items:
                      type: integer
                    example: [1, 2, 3, 4, 5]
        "404":
          description: Unknown exam slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /exams/{examSlug}/questions/{number}:
    parameters:
      - $ref: "#/components/parameters/ExamSlug"
      - name: number
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Question number, stable within the exam
    get:
      summary: Fetch a single question with choices, answer key, explanation
      operationId: getQuestion
      responses:
        "200":
          description: Question payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Unknown exam or question number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    ExamSlug:
      name: examSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]*$"
      example: sap-c02
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
        poolVersion:
          type: string
          example: v2025-11-26
        loadedAt:
          type: string
          format: date-time
    ExamSummary:
      type: object
      required: [slug, displayName, version, totalQuestions]
      properties:
        slug:
          type: string
          example: sap-c02
        displayName:
          type: string
          example: "AWS SAP-C02"
        version:
          type: string
          example: v2025-11-26
        totalQuestions:
          type: integer
          example: 476
    Choice:
      type: object
      required: [label, text]
      properties:
        label:
          type: string
          minLength: 1
          example: A
        text:
          type: string
    Question:
      type: object
      required: [examSlug, number, text, choices, answerKey]
      properties:
        examSlug:
          type: string
          example: sap-c02
        number:
          type: integer
          example: 12
        text:
          type: string
        choices:
          type: array
          minItems: 1
          maxItems: 8
          items:
            $ref: "#/components/schemas/Choice"
        answerKey:
          type: array
          minItems: 1
          items:
            type: string
          description: Subset of `choices[].label`. Length 1 = single-answer; >1 = multi-answer.
        explanation:
          type: string
          nullable: true
    ErrorEnvelope:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - NOT_FOUND
            - VALIDATION
            - POOL_UNAVAILABLE
            - INTERNAL
          example: NOT_FOUND
        message:
          type: string
          example: "해당 문제를 찾을 수 없습니다."
        details:
          type: object
          additionalProperties: true
          nullable: true
        requestId:
          type: string
          example: "9c3a7f12-1d2e-4c3a-b1ad-22f0d0d4a4e1"